🚀CHANGE: Move function tests to subfolder
