+++ tests/test_functions/__init__.py
@@ -1,0 +1,1 @@
+# Almost empty file
